+++ CHANGELOG.md
@@ -15,6 +15,7 @@ SPDX-License-Identifier: GPL-3.0-or-later
 ## [2.5.0](https://git.ufz.de/rdm-software/saqc/-/tags/v2.4.1) - 2023-06-22
 [List of commits](https://git.ufz.de/rdm-software/saqc/-/compare/v2.4.1...v2.5.0)
 ### Added
+- Function selection via strings for most function-expecting parameters
 - `SaQC.plot`:
   - enable multivariate plots
   - keyword `plot_kwargs` to pass matplotlib related arguments